Move all report modules to events.
