+++ .arc
@@ -31,7 +31,6 @@ status      # Status updater
 @scheduled
 runner rate(2 hours)            # Regularly invokes crawls and scrapes
 regen-timeseries rate(2 hours)  # Regularly regenerates timeseries sources
-reports rate(6 hours)           # Generates reports
 
 
 @storage-public
