+++ src/events/reports/generate-data/_build-timeseries.js
