+++ src/events/reports/package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
-    "@architect/functions": "^3.8.4"
+    "@architect/functions": "^3.8.4",
+    "csv-stringify": "^5.5.0"
   }
 }
