--- src/scheduled/reports/.arc-config
@@ -1,4 +1,0 @@
-@aws
-runtime nodejs12.x
-memory 256
-timeout 300
